Fix imports for all uses - click, package, direct run and tests
