+++ further_link/__init__.py
@@ -1,3 +1,3 @@
-from .keyboard_button import KeyboardButton
-from .send_image import send_image
-from .version import __version__
+from further_link.lib.keyboard_button import KeyboardButton
+from further_link.lib.send_image import send_image
+from further_link.version import __version__
