+++ further_link/runner/py_process_handler.py
@@ -3,8 +3,8 @@ import pathlib
 
 import aiofiles
 
-from .process_handler import ProcessHandler
-from .util.user_config import get_absolute_path, get_working_directory
+from further_link.runner.process_handler import ProcessHandler
+from further_link.util.user_config import get_absolute_path, get_working_directory
 
 dirname = pathlib.Path(__file__).parent.absolute()
 further_link_module_path = os.path.join(dirname, "lib")
