+++ tests/e2e/test_http_upload.py
@@ -5,7 +5,7 @@ import aiofiles
 import aiohttp
 import pytest
 
-from further_link.upload import get_bucket_cache_path, get_directory_path
+from further_link.util.upload import get_bucket_cache_path, get_directory_path
 
 from ..dirs import WORKING_DIRECTORY
 from . import UPLOAD_URL
