+++ tests/e2e/test_run_py_upload.py
@@ -4,8 +4,8 @@ import os
 import aiofiles
 import pytest
 
-from further_link.upload import get_bucket_cache_path, get_directory_path
 from further_link.util.message import create_message, parse_message
+from further_link.util.upload import get_bucket_cache_path, get_directory_path
 
 from ..dirs import WORKING_DIRECTORY
 from .helpers import receive_data, wait_for_data
